Refactoring BCCP CUD parts.

+++ src/main/java/org/oagi/srt/gateway/http/api/cc_management/data/CcUpdateManifestRequest.java
@@ -2,8 +2,10 @@ package org.oagi.srt.gateway.http.api.cc_management.data;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class CcUpdateManifestRequest {
-    private long accManifestId;
-    private long bdtManifestId;
+    private BigInteger accManifestId;
+    private BigInteger bdtManifestId;
 }
